+++ app/react/V2/testing/index.ts
@@ -1,3 +1,3 @@
 // eslint-disable-next-line camelcase
-export { LEGACY_createStore } from './testingHelpers';
+export { LEGACY_createStore } from './reduxStore';
 export { TestAtomStoreProvider } from './TestAtomStoreProvider';
